// sanity.types.ts
export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export type Product = {
  _id: string;
  _type: "product";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  slug?: Slug;
  description?: Array<{
    children?: Array<{
      marks?: Array<string>;
      text?: string;
      _type: "span";
      _key: string;
    }>;
    style?: "normal" | "h1" | "h2" | "h3" | "h4" | "h5" | "h6" | "blockquote";
    listItem?: "bullet" | "number";
    markDefs?: Array<{
      href?: string;
      _type: "link";
      _key: string;
    }>;
    level?: number;
    _type: "block";
    _key: string;
  }>;
  price?: number;
  discountPrice?: number;
  discountPercentage?: number;
  images?: Array<{
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
    _key: string;
  }>;
  category?: Category;
  brand?: Brand;
  colorVariants?: Array<{
    color?:
      | "Black"
      | "White"
      | "Red"
      | "Blue"
      | "Green"
      | "Yellow"
      | "Grey"
      | "Pink";
    stock?: number;
    additionalPrice?: number;
    _type: "colorVariant";
    _key: string;
  }>;
  sizeVariants?: Array<{
    sizeLabel?: "XS" | "S" | "M" | "L" | "XL" | "2XL" | "3XL";
    stock?: number;
    additionalPrice?: number;
    _type: "sizeVariant";
    _key: string;
  }>;
  sizeInchesVariants?: Array<{
    sizeInches?: number;
    stock?: number;
    additionalPrice?: number;
    _type: "inchesVariant";
    _key: string;
  }>;
  rating?: number;
  stock?: number;
  isNew?: boolean;
  isFeatured?: boolean;
  totalSales?: number;
  isHotDeal?: boolean;
  isBestSelling?: boolean;
  metaTitle?: string;
  metaDescription?: string;
};

export type Slug = {
  _type: "slug";
  current?: string;
  source?: string;
};

export type Brand = {
  _id: string;
  _type: "brand";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  slug?: Slug;
  description?: string;
  image?: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  };
};
// ...
export type Category = {
  _id: string;
  _type: "category";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  slug?: Slug;
  description?: string;
  featured?: boolean;
  icon?: string;
  image?: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  };
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;